requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
telethon==1.41.2
python-dotenv==1.1.1
pydantic==2.11.9
